+++ client/package.json
@@ -8,20 +8,17 @@
 		"@testing-library/user-event": "^13.5.0",
 		"base-station": "file:..",
 		"chart.js": "^3.7.1",
+		"next": "^12.2.0",
 		"react": "^18.1.0",
 		"react-chartjs-2": "^4.1.0",
 		"react-dom": "^18.1.0",
-		"react-scripts": "5.0.1",
-		"serve": "^13.0.2",
 		"socket.io-client": "^4.5.0",
 		"web-vitals": "^2.1.4"
 	},
 	"scripts": {
-		"start": "react-scripts start",
-		"build": "react-scripts build",
-		"test": "react-scripts test",
-		"eject": "react-scripts eject",
-		"serve": "serve -s build"
+		"dev": "next dev",
+		"build": "next build",
+		"start": "next start"
 	},
 	"eslintConfig": {
 		"extends": [
