+++ client/styles/Home.module.css
@@ -1,0 +1,9 @@
+.container {
+  padding: 0 2rem;
+}
+
+.main {
+  text-align: center;
+}
+
+
